refactor: split config into a config/ subdirectory
config.h/config.cpp had grown to ~920 lines spanning the data model, address-
family policy, validation, the parse/merge engine, and formatting. Split them
into src/reflector/config/:

  address_family.h  AddressFamily enum + Uses/Requires policy + its formatter
  entries.h/.cpp    Wol/Mdns/Ssdp config structs + formatters + Verify()
  config.h          EnvVar + the Config class + its formatter (public header)
  config.cpp        the parse/merge engine + Config method definitions

The subdir's CMakeLists.txt adds its sources to the existing reflector target
(no separate library). Pure code movement, no behavior change.
