+++ src/reflector/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(reflector STATIC
     ${CMAKE_CURRENT_SOURCE_DIR}/default_address_monitor.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/application.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/checksum.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/config.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/dial_proxy.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/dynamic_family_reflector.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/event_loop_dispatcher.cpp
@@ -28,3 +27,5 @@ reflector_configure_project_target(reflector)
 target_include_directories(reflector PUBLIC ${PROJECT_SOURCE_DIR}/src)
 target_link_libraries(reflector PRIVATE tomlplusplus::tomlplusplus)
 target_compile_definitions(reflector PRIVATE TOML_EXCEPTIONS=0)
+
+add_subdirectory(config)
