+++ src/reflector/config/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(reflector PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/entries.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/config.cpp
+)
