+++ src/reflector/mdns_reflector.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "config.h"
+#include "config/config.h"
 #include "dynamic_family_reflector.h"
 #include "ip_address.h"
 #include "link_socket.h"
