+++ tests/config_test.cpp
@@ -1,4 +1,4 @@
-#include "reflector/config.h"
+#include "reflector/config/config.h"
 #include "reflector/logger.h"
 #include "reflector/mac_address.h"
